Test against serde with no features in both std and no-std
