+++ src/datetime.rs
@@ -923,37 +923,12 @@ pub mod rustc_serialize {
 #[cfg(feature = "serde")]
 pub mod serde {
     use core::fmt;
-    // #[cfg(any(test, feature = "alloc"))]
-    // use alloc::format;
     use super::DateTime;
     #[cfg(feature="clock")]
     use offset::Local;
     use offset::{LocalResult, TimeZone, Utc, FixedOffset};
     use serdelib::{ser, de};
-
-    enum SerdeError<V: fmt::Display, D: fmt::Display> {
-        NonExistent { timestamp: V },
-        Ambiguous { timestamp: V, min: D, max: D }
-    }
-
-    impl<V: fmt::Display, D: fmt::Display> fmt::Debug for SerdeError<V, D> {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            write!(f, "ChronoSerdeError({})", self)
-        }
-    }
-
-    // impl<V: fmt::Display, D: fmt::Debug> core::error::Error for SerdeError<V, D> {}
-    impl<V: fmt::Display, D: fmt::Display> fmt::Display for SerdeError<V, D> {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            match self {
-                &SerdeError::NonExistent { ref timestamp } => write!(
-                    f, "value is not a legal timestamp: {}", timestamp),
-                &SerdeError::Ambiguous { ref timestamp, ref min, ref max } => write!(
-                        f, "value is an ambiguous timestamp: {}, could be either of {}, {}",
-                        timestamp, min, max),
-            }
-        }
-    }
+    use {SerdeError, ne_timestamp};
 
     // try!-like function to convert a LocalResult into a serde-ish Result
     fn serde_from<T, E, V>(me: LocalResult<T>, ts: &V) -> Result<T, E>
@@ -964,7 +939,7 @@ pub mod serde {
     {
         match me {
             LocalResult::None => Err(E::custom(
-                SerdeError::NonExistent::<_, u8> { timestamp: ts })),
+                ne_timestamp(ts))),
             LocalResult::Ambiguous(min, max) => Err(E::custom(
                 SerdeError::Ambiguous { timestamp: ts, min: min, max: max })),
             LocalResult::Single(val) => Ok(val)
